GH-139 Hide OAuth2UserDetails service
