+++ reposilite-backend/src/main/java/org/panda_lang/reposilite/authentication/OAuth2UserDetailsFactory.java
@@ -14,22 +14,21 @@
  * limitations under the License.
  */
 
-package org.panda_lang.reposilite.authentication.userdetails;
+package org.panda_lang.reposilite.authentication;
 
 import java.util.Map;
 
-public interface OAuth2UserDetails {
+public final class OAuth2UserDetailsFactory {
 
-    String getProviderId();
+    private OAuth2UserDetailsFactory() {}
 
-    String getName();
-
-    String getDisplayName();
-
-    String getEmail();
-
-    String getAvatar();
-
-    Map<String, Object> getAttributes();
+    public static OAuth2UserDetails obtainOAuth2UserDetails(String providerName, Map<String, Object> attributes) {
+        if ("github".equalsIgnoreCase(providerName)) {
+            return new GithubOAuth2UserDetails(attributes);
+        }
+        else {
+            throw new UnsupportedOperationException();
+        }
+    }
 
 }
