+++ reposilite-backend/src/main/java/org/panda_lang/reposilite/authentication/OAuth2UserDetailsService.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package org.panda_lang.reposilite.authentication.userdetails;
+package org.panda_lang.reposilite.authentication;
 
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 
-public abstract class OAuth2UserDetailsService extends DefaultOAuth2UserService {
+abstract class OAuth2UserDetailsService extends DefaultOAuth2UserService {
 
 }
